refactor: ABC/MIDI/MusicXML/Verovio のI/O責務を `main.ts` から分離 ### 概要 `main.ts` に集中していた I/O/変換処理をモジュール分割し、DOMベースのインタフェースを導入しました。 あわせて ABC 関連ファイルを `abc-io.ts` に統合しています。
### 主な変更
1. ABC関連の統合
- `src/ts/abc-common.ts` を削除
- `src/ts/abc-compat-parser.ts` を `src/ts/abc-io.ts` に統合（実質リネーム+集約）
- `main.ts` 側 import を `abc-io.ts` に統一

2. MIDI処理の分離
- `src/ts/midi-io.ts` を新規追加
- `buildPlaybackEventsFromMusicXmlDoc(doc, ticks)` を追加（DOM入力）
- `buildMidiBytesForPlayback(...)` を `midi-io.ts` 側へ集約
- `src/ts/playback.ts` は再エクスポート層へ簡素化
- `main.ts` は MIDI処理を `midi-io.ts` 経由に変更

3. MusicXML I/Oの分離（DOMインタフェース）
- `src/ts/musicxml-io.ts` を新規追加
- 追加API:
  - `parseMusicXmlDocument`
  - `serializeMusicXmlDocument`
  - `buildRenderDocWithNodeIds`
  - `extractMeasureEditorDocument`
  - `replaceMeasureInMainDocument`
- `main.ts` 内の小節抽出/差し替え/描画前処理を置換

4. Verovio描画処理の分離
- `src/ts/verovio-out.ts` を新規追加
- `renderMusicXmlDomToSvg(doc, options)` を追加（DOM入力）
- `main.ts` から verovio 初期化・描画詳細を切り離し

5. ビルド成果物更新
- `src/js/main.js`, `mikuscore.html`, `mikuscore-src.html` を更新

### 影響範囲
- 主に内部構造のリファクタリング
- 既存機能のUI仕様変更は最小（処理経路をモジュール化）

### 目的/効果
- `main.ts` の肥大化抑制
- 機能単位での責務明確化
- DOMベースAPIによる後続拡張（検証・差分適用・描画経路の共通化）を容易化
