+++ mikuscore-src.html
@@ -85,14 +85,6 @@
               </label>
               <label class="ms-field">調号
                 <select id="newKeyFifths" class="md-select">
-                  <option value="-7">♭7 (Cb)</option>
-                  <option value="-6">♭6 (Gb)</option>
-                  <option value="-5">♭5 (Db)</option>
-                  <option value="-4">♭4 (Ab)</option>
-                  <option value="-3">♭3 (Eb)</option>
-                  <option value="-2">♭2 (Bb)</option>
-                  <option value="-1">♭1 (F)</option>
-                  <option value="0" selected>なし (C)</option>
                   <option value="1">♯1 (G)</option>
                   <option value="2">♯2 (D)</option>
                   <option value="3">♯3 (A)</option>
@@ -100,6 +92,14 @@
                   <option value="5">♯5 (B)</option>
                   <option value="6">♯6 (F#)</option>
                   <option value="7">♯7 (C#)</option>
+                  <option value="0" selected>なし (C)</option>
+                  <option value="-1">♭1 (F)</option>
+                  <option value="-2">♭2 (Bb)</option>
+                  <option value="-3">♭3 (Eb)</option>
+                  <option value="-4">♭4 (Ab)</option>
+                  <option value="-5">♭5 (Db)</option>
+                  <option value="-6">♭6 (Gb)</option>
+                  <option value="-7">♭7 (Cb)</option>
                 </select>
               </label>
             </div>
@@ -144,7 +144,7 @@
 
         <section class="md-section ms-flow-step">
           <h2 class="md-section-title">
-            <span class="ms-step-no">2</span>譜面プレビュー（クリックで選択）
+            <span class="ms-step-no">2</span>フルスコア
             <span class="md-tooltip-group">
               <span class="md-info-chip" aria-label="譜面プレビューの説明">
                 <svg aria-hidden="true" viewBox="0 0 24 24" class="md-info-icon" fill="none">
